blobs fetcher unification
