+++ ydb/core/tx/columnshard/engines/reader/common_reader/iterator/ya.make
@@ -1,13 +1,14 @@
 LIBRARY()
 
 SRCS(
-    fetched_data.cpp
     columns_set.cpp
-    iterator.cpp
+    constructor.cpp
     context.cpp
-    source.cpp
-    fetching.cpp
     fetch_steps.cpp
+    fetched_data.cpp
+    fetching.cpp
+    iterator.cpp
+    source.cpp
 )
 
 PEERDIR(
