+++ ydb/core/tx/columnshard/engines/reader/plain_reader/iterator/ya.make
@@ -2,7 +2,6 @@ LIBRARY()
 
 SRCS(
     scanner.cpp
-    constructor.cpp
     source.cpp
     interval.cpp
     fetched_data.cpp
